Add platform abstraction for term and pty
