+++ rust_pty/Cargo.toml
@@ -4,5 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-nix = { version = "0.27", features = ["term"] }
 tokio = { version = "1", features = ["rt", "net", "time", "macros", "io-util"] }
+
+[target.'cfg(unix)'.dependencies]
+nix = { version = "0.27", features = ["term"] }
