+++ rust_term/Cargo.toml
@@ -4,4 +4,4 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-libc = "0.2"
+nix = { version = "0.27", features = ["term", "ioctl"] }
