+++ rust_term/src/platform/mod.rs
@@ -1,0 +1,9 @@
+#[cfg(unix)]
+mod unix;
+#[cfg(unix)]
+pub use unix::*;
+
+#[cfg(not(unix))]
+mod fallback;
+#[cfg(not(unix))]
+pub use fallback::*;
